@@ -1,10 +1,10 @@
 cmake_minimum_required(VERSION 3.16)
 project(Scorpion)
 
-set(CMAKE_CXX_STANDARD 14)
+set(CMAKE_CXX_STANDARD 17)
 set(CMAKE_CXX_FLAGS "-Wall -Wextra -Wconversion -Wsizeof-pointer-memaccess \
-                     -Wfloat-equal -Wconversion-null -Woverflow -Wshadow \
-                     -D_FILE_OFFSET_BITS=64 -D_GNU_SOURCE -pthread -g -O3")
+                     -Wfloat-equal -Wconversion-null -Woverflow -Wshadow -faligned-new\
+                     -D_FILE_OFFSET_BITS=64 -D_GNU_SOURCE -pthread -g -O0")
 
 include_directories(
         src/algorithm
@@ -30,6 +30,7 @@ foreach (_target
         LockFreeQueue
         NRWLock
         SpinLockMutex
+        ThreadPool
         TimeWheel
         TokenBucket
         UnixSocket
