@@ -24,7 +24,7 @@ public:
     TimeWheelRaw &operator=(const TimeWheelRaw &) = delete;
 
 public:
-    void Add(std::function<void()> &&cb, unsigned int interval, int loop);
+    void Add(std::function<int()> &&cb, unsigned int interval, int loop);
     void Tick();
     void Dump() const;
 
@@ -57,9 +57,9 @@ public:
     TimeWheel &operator=(const TimeWheel &) = delete;
 
 public:
-    void Add(std::function<void()> &&cb, unsigned int interval, int loop) {
+    void Add(std::function<int()> &&cb, unsigned int interval, int loop) {
         std::lock_guard<std::mutex> lock(_mutex);
-        _twr->Add(std::forward<std::function<void()>>(cb), interval, loop);
+        _twr->Add(std::forward<std::function<int()>>(cb), interval, loop);
     }
 
 private:
